+++ src/components/revision_files.rs
@@ -1,11 +1,12 @@
 use super::{
-	utils::scroll_vertical::VerticalScroll, BlameFileOpen,
-	CommandBlocking, CommandInfo, Component, DrawableComponent,
-	EventState, FileRevOpen, FuzzyFinderTarget, SyntaxTextComponent,
+	utils::scroll_vertical::VerticalScroll, CommandBlocking,
+	CommandInfo, Component, DrawableComponent, EventState,
+	FuzzyFinderTarget, SyntaxTextComponent,
 };
 use crate::{
 	app::Environment,
 	keys::{key_match, SharedKeyConfig},
+	popups::{BlameFileOpen, FileRevOpen},
 	queue::{InternalEvent, Queue, StackablePopupOpen},
 	strings::{self, order, symbol},
 	try_or_popup,