+++ src/popups/confirm.rs
@@ -1,8 +1,8 @@
 use crate::{
 	app::Environment,
 	components::{
-		popup_paragraph, visibility_blocking, CommandBlocking,
-		CommandInfo, Component, DrawableComponent, EventState,
+		visibility_blocking, CommandBlocking, CommandInfo, Component,
+		DrawableComponent, EventState,
 	},
 	keys::{key_match, SharedKeyConfig},
 	queue::{Action, InternalEvent, Queue},
@@ -16,8 +16,10 @@ use ratatui::{
 use std::borrow::Cow;
 use ui::style::SharedTheme;
 
+use super::popup_paragraph;
+
 ///
-pub struct ConfirmComponent {
+pub struct ConfirmPopup {
 	target: Option<Action>,
 	visible: bool,
 	queue: Queue,
@@ -25,7 +27,7 @@ pub struct ConfirmComponent {
 	key_config: SharedKeyConfig,
 }
 
-impl DrawableComponent for ConfirmComponent {
+impl DrawableComponent for ConfirmPopup {
 	fn draw<B: Backend>(
 		&self,
 		f: &mut Frame<B>,
@@ -51,7 +53,7 @@ impl DrawableComponent for ConfirmComponent {
 	}
 }
 
-impl Component for ConfirmComponent {
+impl Component for ConfirmPopup {
 	fn commands(
 		&self,
 		out: &mut Vec<CommandInfo>,
@@ -102,7 +104,7 @@ impl Component for ConfirmComponent {
 	}
 }
 
-impl ConfirmComponent {
+impl ConfirmPopup {
 	///
 	pub fn new(env: &Environment) -> Self {
 		Self {