+++ src/popups/log_search.rs
@@ -1,6 +1,6 @@
-use super::{
+use crate::components::{
 	visibility_blocking, CommandBlocking, CommandInfo, Component,
-	DrawableComponent, EventState, TextInputComponent,
+	DrawableComponent, EventState, InputType, TextInputComponent,
 };
 use crate::{
 	app::Environment,
@@ -41,7 +41,7 @@ enum PopupMode {
 	JumpCommitSha,
 }
 
-pub struct LogSearchPopupComponent {
+pub struct LogSearchPopupPopup {
 	repo: RepoPathRef,
 	queue: Queue,
 	visible: bool,
@@ -54,12 +54,12 @@ pub struct LogSearchPopupComponent {
 	jump_commit_id: Option<CommitId>,
 }
 
-impl LogSearchPopupComponent {
+impl LogSearchPopupPopup {
 	///
 	pub fn new(env: &Environment) -> Self {
 		let mut find_text =
 			TextInputComponent::new(env, "", "search text", false)
-				.with_input_type(super::InputType::Singleline);
+				.with_input_type(InputType::Singleline);
 		find_text.embed();
 		find_text.enabled(true);
 
@@ -524,7 +524,7 @@ impl LogSearchPopupComponent {
 	}
 }
 
-impl DrawableComponent for LogSearchPopupComponent {
+impl DrawableComponent for LogSearchPopupPopup {
 	fn draw<B: Backend>(
 		&self,
 		f: &mut Frame<B>,
@@ -545,7 +545,7 @@ impl DrawableComponent for LogSearchPopupComponent {
 	}
 }
 
-impl Component for LogSearchPopupComponent {
+impl Component for LogSearchPopupPopup {
 	fn commands(
 		&self,
 		out: &mut Vec<CommandInfo>,