+++ src/popups/stashmsg.rs
@@ -1,7 +1,6 @@
-use super::{
-	textinput::TextInputComponent, visibility_blocking,
-	CommandBlocking, CommandInfo, Component, DrawableComponent,
-	EventState,
+use crate::components::{
+	visibility_blocking, CommandBlocking, CommandInfo, Component,
+	DrawableComponent, EventState, InputType, TextInputComponent,
 };
 use crate::{
 	app::Environment,
@@ -15,7 +14,7 @@ use asyncgit::sync::{self, RepoPathRef};
 use crossterm::event::Event;
 use ratatui::{backend::Backend, layout::Rect, Frame};
 
-pub struct StashMsgComponent {
+pub struct StashMsgPopup {
 	repo: RepoPathRef,
 	options: StashingOptions,
 	input: TextInputComponent,
@@ -23,7 +22,7 @@ pub struct StashMsgComponent {
 	key_config: SharedKeyConfig,
 }
 
-impl DrawableComponent for StashMsgComponent {
+impl DrawableComponent for StashMsgPopup {
 	fn draw<B: Backend>(
 		&self,
 		f: &mut Frame<B>,
@@ -35,7 +34,7 @@ impl DrawableComponent for StashMsgComponent {
 	}
 }
 
-impl Component for StashMsgComponent {
+impl Component for StashMsgPopup {
 	fn commands(
 		&self,
 		out: &mut Vec<CommandInfo>,
@@ -124,7 +123,7 @@ impl Component for StashMsgComponent {
 	}
 }
 
-impl StashMsgComponent {
+impl StashMsgPopup {
 	///
 	pub fn new(env: &Environment) -> Self {
 		Self {
@@ -136,7 +135,7 @@ impl StashMsgComponent {
 				&strings::stash_popup_msg(&env.key_config),
 				true,
 			)
-			.with_input_type(super::InputType::Singleline),
+			.with_input_type(InputType::Singleline),
 			key_config: env.key_config.clone(),
 			repo: env.repo.clone(),
 		}